Starting with online multiplayer functionality and update menu
The work on the Online multiplayer functionality has been started through the addition of the NetworkingHandler class, allowing for local and online gameplay. The menu has been updated to accommodate this new feature, offering options to play locally, connect online, or create an online game. Reactivity of the menu options with mouse interactions has been introduced.